+++ src/Integer_Complement.cpp
@@ -5,7 +5,7 @@ InfiniteArithmetic::Integer InfiniteArithmetic::Integer::Complement() const
     Integer complement;
     complement = *this;
 
-    for(int i=0; i<Size; i++)
+    for(size_t i=0; i<Size; i++)
     {
         complement.Array[i] = 9 - complement.Array[i];
     }
